Fixes build properties for win32 release.
The old properties were correct for Debug but not Release. The build
system needs to build object files in subdirectories because of name
conflicts. So, I just copied and pasted the value for Object file
directory to contain some special token for creating subdirectories. It
was super frustrating to find in the first place. Ugh.