Removes commit method from Actor.
Unused method. Object is mutable from any updater, so conditions will in
fact stack. This is fine because updaters should not be doing too much
comparing and events that affect the player will be performed in a
different area.

This means frames are not transactions. Maybe it will become more
reasonable, but the amount of complexity added would be too much. So,
the order of actions might matter, but that order is determined by the
game designer.

It probably would be either immediately noticable in the case that the
conditions are common, or very obscure in which the problem will be
rather curious but not horrible. Although this means bugs triggered by
ordering will be hard to reproduce.