Adds Engine::Object to handle pathing and object loading.
This Engine will coordinate with the individual Registry::* classes and
will load various objects in the system. This Engine understands the
path and file system hierarchy given to it and can easily figure out
where objects should be loaded from.

The Object Engine will easily be able to read generic object
descriptions and will then pass on relevant parts of larger JSON
descriptions to individual Registry constructors. This should allow
better flexibility for project descriptions.

Another bonus to add to the Object Engine is the ability to compress
game files. This Engine will be responsible for decompressing.