Enforces Object engine creation with System; JSON.
Adds JSON parsing for object engine; the JSON indicates the paths
various objects for the system are found. When other parts of the
system refer to an existing object by name, the object engine can then
know how to find that object's description and load it if it is absent.

Still need to add loaders for various types.

Still need to document final specifications for the object JSON layouts.