Optimizes MapCollider rule.
Instead of checking against every tile of a map, just check for
collisions of the tiles that are known to overlap with the actor. Now
collisions can scale with the map size.

It is unknown why I need to add 2 to the max_x/y for this to work. It
makes little sense. I should look into that. But it is better than
checking every single tile.