Adds Registry::Scene.
This class represents a Scene that has not been instantiated. This class
is necessary since Scenes may be reinitialized. For instance, in a game,
a scene can be a level. If the player fails at some task, the scene may
be restarted.

Therefore, this Scene class is the result of loading a JSON description.
It is immutable. It can be used to construct a World::Scene that
represents a mutable, instance of a Scene that a person or machine can
interact with.

I am now starting to pass the Object engine to individual object loaders
so that they can also load objects with the pathing given for the
System. It has created lots of cycle problems, but they are easily
solved.