Adds Response event support.
It can now respond to collision events. Yet, I need to set whether or
not a collision impedes movement. And I need to figure out how to handle
'hitting' an actor and 'being on' an actor. Right now, if you run into
an actor, it will trigger a collision every frame. It should probably
remember that it collided last frame.