Adds Sprite::Font - reader, renderer, cacher of fonts.
Right now, the Sprite::Font class will lazily render individual glyphs
of a set size. So, I have to add size adjustment. I would also like to
tear apart this work into a separate tool so that FreeType is not a
requirement to execute a game.

Also, in the future, I will extract curve data and render fonts in
real-time as well as bitmaps for smaller sizes. Only small sizes must
have bitmap data. The tool will extract this out and allow it to be used
in any Engine::System implementation in any language without having to
port FreeType.