Fixes font aliasing issues.
Since fonts are rendered with luminosity values instead of color, the
correct thing to do is use the luminosity as the alpha value and make
the color white. Then it will both blend correctly and be affected by
color through fragment shaders the correct way.

Tutorials for combining FreeType and OpenGL are horribly flawed in this
manner. They suggest averaging the luminosity to create the color. That
only works if you assume your background color is 0 (black). Come on,
Internet.