Adds Audio::Music to hold music data.
Engine::Audio now plays the music it is given.

Next: To handle looping and callbacks for music finishing.

Afterward: To handle system events in a manner that can be used in
objects. For instance, the music ending should trigger an Event that can
be placed on an object! That'd be cool at least.