Adds metadata for mouse position on mouse input events and Font metrics.
It now routes mouse presses to the Window underneath the cursor,
although the logic is iterating through the windows the wrong order.
Now, we just need to add event functions for handling input.

Sprite::Font can now return correct font metrics and can determine the
size of a given string. The text for Button is now centered.

Adds a little bit of fleshing out to Sprite::Layout.